Added champion post static route, test is passing
