+++ lib/app.js
@@ -4,7 +4,7 @@ import errorMiddleware from './middleware/error.js';
 import bunniesController from './controllers/bunnies.js';
 import sundaesController from './controllers/sundaes.js';
 import charactersController from './controllers/characters.js';
-import championsController from './champions/characters.js';
+import championsController from './controllers/champions.js';
 
 const app = express();
