+++ lib/models/Champion.js
@@ -1,28 +1,20 @@
-import pool from '../lib/utils/pool.js';
-import setup from '../data/setup.js';
-import request from 'supertest';
-import app from '../lib/app.js';
-import Champion from '../lib/models/Champion.js';
+import pool from "../utils/pool";
 
-describe('champion routes', () => {
-  beforeEach(() => {
-    return setup(pool);
-  });
-  
-  it('creates a champion via POST', async () => {
-    const lux = { 
-      name: 'Luxanna Crownguard', 
-      role: 'mage', 
-      damage: 'ability power', 
-      difficulty: 2,
-      region: 'Demacia' };
-    const res = await request(app)
-      .post('/api/v1/bunnies')
-      .send(lux);
-  
-    expect(res.body).toEqual({
-      id: '1',
-      ...lux
-    });
-  });
-});
+export default class Champion {
+    id;
+    name;
+    role;
+    damage;
+    difficulty;
+    region;
+
+    constructor(row) {
+        this.id = row.id;
+        this.name = row.name;
+        this.role = row.role;
+        this.damage = row.damage;
+        this.difficulty = row.difficulty;
+        this.region = row.region;
+    }
+}
+
